Clear completed task implementation done
